Add .gitignore, add necessary folders
When working with a node project we want to avoid uploading node_modules because anyone can have their node project install these with npm install.
Likewise we don't need to upload the dist folder because it holds what is created by webpack and again anyone that runs webpack in this project will build the same dist.
I forgot to add vs code settings which I'll have to do in the near future to be ignored.