Add starting modules for node server
We use 3 different webpack modules in this project and they are included here for the basic use of webpack, giving webpack cli commands, and being able to have our server update when we change any javascript files on save.
We also use a module that is used by webpack to read html files and pack them into our bundled result.