Add content to template.html, html-loader
We've added a very small amount of content to our html template for now, and to be able to use this newly added content in the html file we added the html-loader module which will parse the html file data and move it to the result html page. It will also relate the image file used in the source html template file so that the html file in the dist directory will be able to use it as well. Of course to use that module we needed to install it into our modules for the project, and tell webpack how to use it in its config file.