Move content to htmlContent.js
Rather than start our template.html with data we'll create these elements and give them data in a javascript file which we'll import into our entry point file and invoke the function from that content file to populate the html page.
To do this we also added a way for image files to be parse and used with a mode.rule which kept an eye out on 5 different types of images.