Add loaders, and watch tool to config file
Although I probably won't use a css file I have now set up the functionality to import a css file into the entry point to be read by webpack.
I've also added the ability for webpack to watch our html template file for changes so that it can rebundle if a change is made to that file.
And finally added eval-source-map so that we can correctly see the projects sources in chrome dev tools.